requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
pydantic
python-dotenv
pydantic[email]